requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Machine Learning & Embeddings
sentence-transformers>=2.3.0
faiss-cpu==1.7.4

# Utilities
python-dotenv==1.0.0
